Fix remaining PMD errors
1. Merge if statements
2. Remove unused imports
3. Rename an equals method that wasn't *the* equals method

+++ src/org/daisy/dotify/formatter/impl/engine/LayoutEngineImpl.java
@@ -111,10 +111,8 @@ class LayoutEngineImpl implements FormatterEngine {
 				throw new LayoutEngineException("FormatterException while running task.", e);
 			}
 		} finally {
-			if (f != null) {
-				if (!f.delete()) {
-					f.deleteOnExit();
-				}
+			if (f != null && !f.delete()) {
+				f.deleteOnExit();
 			}
 		}
 	}
